Move get_mock_asf_cookie into api conftest
