+++ apps/api/src/hyp3_api/auth.py
@@ -1,4 +1,3 @@
-import time
 from os import environ
 
 import jwt
@@ -27,17 +26,3 @@ def decode_asf_cookie(cookie: str) -> tuple[str, str]:
 
 def get_jwks_client() -> jwt.PyJWKClient:
     return jwt.PyJWKClient('https://urs.earthdata.nasa.gov/.well-known/edl_ops_jwks.json')
-
-
-def get_mock_jwt_cookie(user: str, lifetime_in_seconds: int, access_token: str) -> str:
-    payload = {
-        'urs-user-id': user,
-        'urs-access-token': access_token,
-        'exp': int(time.time()) + lifetime_in_seconds,
-    }
-    value = jwt.encode(
-        payload=payload,
-        key=environ['AUTH_PUBLIC_KEY'],
-        algorithm=environ['AUTH_ALGORITHM'],
-    )
-    return value
